add error handling wrapper
This is pure black magic. Do not tamper with unless you are
prepared to take responsibility for breaking everything.

Essentially, a hidden class overrides the usual class, allowing
methods and attributes to be accessed normally. If a non-blacklisted
method is accessed, this is overridden with wrap_func, which scans
the return dictionary of the usual method for the expected error
response. If the error response is found, the error class is generated
and returned.

Like I said, black magic. run away.